Misc fixes tweaks, also ran code cleanup/sorting.
Fixes:
Fixes some issues with trainers (Battle issues, naming issues)
Fixes some backend issues with choose first gui
Fixes pokecubes not firing from dispensers.
Fixes eggs not placing from dispensers.
Fixes eggs not going into hoppers.

Tweaks:
Added ability to set lists for trainers to drop.
Added ability to set texture for professor.
Added message warning what pokemob the trainer will send out next.
Added configs for trainer agression delays.

Made NBTEdit send an entity update packet to players.

Optimizations for trainer texture code.

+++ Pokecube Core/src/main/java/pokecube/core/client/render/entity/RenderPokemob.java
@@ -113,13 +113,13 @@ public class RenderPokemob<T extends EntityLiving> extends RenderPokemobInfos<T>
                 worldrenderer.pos(0.0D, 0.0D, 0.0D)
                         .color(col1.getRed(), col1.getGreen(), col1.getBlue(), (int) (255.0F * (1.0F - f1)))
                         .endVertex();
-                worldrenderer.pos(-0.866D * (double) f3, (double) f4, (double) (-0.5F * f3))
+                worldrenderer.pos(-0.866D * f3, f4, -0.5F * f3)
                         .color(col2.getRed(), col2.getGreen(), col2.getBlue(), 0).endVertex();
-                worldrenderer.pos(0.866D * (double) f3, (double) f4, (double) (-0.5F * f3))
+                worldrenderer.pos(0.866D * f3, f4, -0.5F * f3)
                         .color(col2.getRed(), col2.getGreen(), col2.getBlue(), 0).endVertex();
-                worldrenderer.pos(0.0D, (double) f4, (double) (1.0F * f3))
+                worldrenderer.pos(0.0D, f4, 1.0F * f3)
                         .color(col2.getRed(), col2.getGreen(), col2.getBlue(), 0).endVertex();
-                worldrenderer.pos(-0.866D * (double) f3, (double) f4, (double) (-0.5F * f3))
+                worldrenderer.pos(-0.866D * f3, f4, -0.5F * f3)
                         .color(col2.getRed(), col2.getGreen(), col2.getBlue(), 0).endVertex();
                 tessellator.draw();
             }
@@ -235,13 +235,13 @@ public class RenderPokemob<T extends EntityLiving> extends RenderPokemobInfos<T>
             worldrenderer.begin(6, DefaultVertexFormats.POSITION_COLOR);
             worldrenderer.pos(0.0D, 0.0D, 0.0D)
                     .color(col1.getRed(), col1.getGreen(), col1.getBlue(), (int) (255.0F * (1.0F - f1))).endVertex();
-            worldrenderer.pos(-0.866D * (double) f3, (double) f4, (double) (-0.5F * f3))
+            worldrenderer.pos(-0.866D * f3, f4, -0.5F * f3)
                     .color(col2.getRed(), col2.getGreen(), col2.getBlue(), 0).endVertex();
-            worldrenderer.pos(0.866D * (double) f3, (double) f4, (double) (-0.5F * f3))
+            worldrenderer.pos(0.866D * f3, f4, -0.5F * f3)
                     .color(col2.getRed(), col2.getGreen(), col2.getBlue(), 0).endVertex();
-            worldrenderer.pos(0.0D, (double) f4, (double) (1.0F * f3))
+            worldrenderer.pos(0.0D, f4, 1.0F * f3)
                     .color(col2.getRed(), col2.getGreen(), col2.getBlue(), 0).endVertex();
-            worldrenderer.pos(-0.866D * (double) f3, (double) f4, (double) (-0.5F * f3))
+            worldrenderer.pos(-0.866D * f3, f4, -0.5F * f3)
                     .color(col2.getRed(), col2.getGreen(), col2.getBlue(), 0).endVertex();
             tessellator.draw();
         }
